Move around animation toolbar items
