+++ src/renderer/components/animator/CaptureButtonToolbar/CaptureButtonToolbar.css
@@ -1,0 +1,5 @@
+.capture-button-toolbar {
+  justify-content: center;
+  padding-bottom: 0;
+  height: var(--height-300);
+}
